backend/requirements.txt
django>=5.1.6
django-jazzmin==3.0.1
fastapi[standard]>=0.115.11
pillow==11.1.0
envist>=0.0.4
pymysql==1.1.1 # For mysql
#psycopg2-2.9.10  # for Postgres
#django-unfold==0.53.0 # Not required
django-import-export[all]>=4.3.7
django-cors-headers==4.7.0
django-ckeditor==6.7.2